♻️ mindmap 분리 및 리팩토링 #20
이름 mindmap lower로 통일
상태 IMindmap -> IMindmapData
molecules / organisms / template 단계 분리
