+++ client/src/components/organisms/index.tsx
@@ -3,3 +3,4 @@ export { default as LoginWrapper } from 'components/organisms/LoginWrapper';
 export { default as NewProjectCard } from 'components/organisms/NewProjectCard';
 export { default as ProjectCard } from 'components/organisms/ProjectCard';
 export { default as TextInputModal } from 'components/organisms/TextInputModal';
+export { default as MindmapBtnWrapper } from 'components/organisms/MindmapBtnWrapper';
