+++ client/src/components/templates/index.tsx
@@ -1,2 +1,3 @@
 export { default as ProjectCardContainer } from 'components/templates/ProjectCardContainer';
 export { default as NewProjectModalWrapper } from 'components/templates/NewProjectModalWrapper';
+export { default as Mindmap } from 'components/templates/Mindmap';
